requirements.txt
streamlit>=1.32.0
playwright>=1.43.0
PyGithub>=2.3.0
ollama>=0.2.0
python-dotenv>=1.0.0
